fix some type error on msvc
using self def type instead if intXX_t

+++ Engine/Z/Core/Core.h
@@ -61,8 +61,8 @@ namespace Z {
     using int16 = short;
     using uint32 = unsigned int;
     using int32 = int;
-    using int64 = long int;
-    using uint64 = unsigned long int;
+    using int64 = long long;
+    using uint64 = unsigned long long;
     static_assert(sizeof(uint8) == 1);
     static_assert(sizeof(uint16) == 2 * sizeof(uint8));
     static_assert(sizeof(int16) == 2 * sizeof(uint8));
